# Conflicts:
#	src/main/java/ru/skypro/homework/repository/AdRepository.java
#	src/main/java/ru/skypro/homework/service/AdService.java
#	src/main/java/ru/skypro/homework/service/impl/AdServiceImpl.java
#	src/main/resources/application.properties

+++ src/main/java/ru/skypro/homework/controller/CommentController.java
@@ -1,70 +1,77 @@
 package ru.skypro.homework.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import ru.skypro.homework.dto.CommentDto;
 import ru.skypro.homework.dto.CommentsDto;
 import ru.skypro.homework.dto.CreateOrUpdateCommentDto;
+import ru.skypro.homework.entity.Ad;
+import ru.skypro.homework.service.AdService;
 import ru.skypro.homework.service.CommentService;
 
+import java.util.Optional;
 
+
+@Slf4j
+@CrossOrigin(value = "http://localhost:3000")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/ads")
+@Tag( name = "Комментарии")
 public class CommentController {
     private CommentService commentService;
+    private AdService adService;
 
 
-    @Operation(summary = "Получение комментариев объявления",
-            responses = {
-                    @ApiResponse(responseCode = "200", content = @Content),
-                    @ApiResponse(responseCode = "401", content = @Content),
-                    @ApiResponse(responseCode = "404", content = @Content)})
-    @GetMapping("/{id}/comments")
-    public ResponseEntity<CommentsDto> getComments(@PathVariable("id") Integer id) {
-        return ResponseEntity.ok(commentService.getComments(id));
+    @Operation(summary = "Получение комментариев объявления")
+    public ResponseEntity<CommentsDto> getComments(@PathVariable Integer id) {
+        CommentsDto comments = commentService.getComments(id);
+        if (comments == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.ok(comments);
     }
 
-    @Operation(summary = "Добавление комментария к объявлению",
-            responses = {
-                    @ApiResponse(responseCode = "200", content = @Content),
-                    @ApiResponse(responseCode = "401", content = @Content),
-                    @ApiResponse(responseCode = "404", content = @Content)})
-    @PostMapping("/{id}/comments")
-    public ResponseEntity<CommentDto> addComment(@PathVariable("id") Integer id,
-                                                 @RequestBody CreateOrUpdateCommentDto createCommentDto) {
-        return new ResponseEntity<>(commentService.createComment(id, createCommentDto), HttpStatus.CREATED);
+
+    @Operation(summary = "Добавление комментария к объявлению")
+    public ResponseEntity<CreateOrUpdateCommentDto> createComment(@PathVariable Integer id, @RequestBody CreateOrUpdateCommentDto createCommentDto, Authentication authentication) {
+        Optional<Ad> ad = adService.findOne(id);
+        if (ad == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        CreateOrUpdateCommentDto commentDto = commentService.createComment(id, createCommentDto, authentication);
+        return ResponseEntity.ok(commentDto);
     }
 
-    @Operation(summary = "Удаление комментария",
-            responses = {
-                    @ApiResponse(responseCode = "200", content = @Content),
-                    @ApiResponse(responseCode = "401", content = @Content),
-                    @ApiResponse(responseCode = "403", content = @Content),
-                    @ApiResponse(responseCode = "404", content = @Content)})
+
+    @Operation(summary = "Удаление комментария")
     @DeleteMapping("/{adId}/comments/{commentId}")
-    public ResponseEntity<Void> deleteComment(@PathVariable("adId") Integer adId,
-                                              @PathVariable("commentId") Integer commentId) {
-
+    public ResponseEntity<CommentDto> deleteComment(@PathVariable Integer adId,
+                                                    @PathVariable Integer commentId) {
+        Optional<Ad> ad = adService.findOne(adId);
+        if (ad == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
         commentService.deleteComment(adId, commentId);
         return ResponseEntity.ok().build();
     }
 
-    @Operation(summary = "Обновление комментария",
-            responses = {
-                    @ApiResponse(responseCode = "200", content = @Content),
-                    @ApiResponse(responseCode = "401", content = @Content),
-                    @ApiResponse(responseCode = "403", content = @Content),
-                    @ApiResponse(responseCode = "404", content = @Content)})
+
+    @Operation(summary = "Обновление комментария")
     @PatchMapping("/{adId}/comments/{commentId}")
-    public ResponseEntity<CommentDto> updateComment(@PathVariable("adId") Integer adId,
-                                                    @PathVariable("commentId") Integer commentId,
-                                                    @RequestBody CreateOrUpdateCommentDto updateCommentDto) {
-        return ResponseEntity.ok(commentService.updateComment(adId, commentId, updateCommentDto));
+    public ResponseEntity<CreateOrUpdateCommentDto> updateComment (@PathVariable Integer adId,
+                                                                   @PathVariable Integer commentId,@RequestBody CreateOrUpdateCommentDto updateComment){
+        CreateOrUpdateCommentDto comment =  commentService.updateComment(adId, commentId, updateComment);
+        if (comment == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+        return ResponseEntity.ok(comment);
     }
+
 }
